Cgroups addition for process max
Mounting is stil on hold. We need Ubuntu filesystem for that.

CGroups do not work - as far as I am aware - with rootless containers.
For now, running as a root user 'disarms' the fork bomb.

`main.go` in this setting is the backbone of the container. There is
nothing that I have added now.