Somehow initialized
Here, we built on Liz Rice's sequel. I broadly described what I am going
to do. From here on, I 'take on', will refactor the code, and add some
tests only so that we see this backbone works. After test inclusions,
containerization will not be worked on until the image part is done.

The project architecture is NOT established yet. After the design
processes finishes, and the project frame is visible, then it will be
decided.