Segregating the documentation from code
One of the most important parts is unraveling the mystery of containers
and images. Now, the code does not contribute to anything for us to understand. I segregated this part for this purpose. Without getting to a certain point in this branch, other branches will wait. This branch not only will include the README and TODOs but also the deep diving all the processes that we will go through. Like a surgeon, I plan to operate on todays technologies and how/what they are doing with the best resolution I can get into. After we have a deep understanding of the underlying the technologies, every code will make sense. So I think this branch is to be the starting point.