Container documentation (#2)
* The bare back bone of the container

go run main.go run <cmd> [args]

Now, if you are a root user you can directly execute any command. If not
you should sudo it first. There is nothing here now that isolates the
process from the host except the Unix Timesharing System which creates
an isolation around the hostname only.

* Cgroups addition for process max

Mounting is stil on hold. We need Ubuntu filesystem for that.

CGroups do not work - as far as I am aware - with rootless containers.
For now, running as a root user 'disarms' the fork bomb.

`main.go` in this setting is the backbone of the container. There is
nothing that I have added now.

* Cgroup SS file name correction

* Somehow initialized

Here, we built on Liz Rice's sequel. I broadly described what I am going
to do. From here on, I 'take on', will refactor the code, and add some
tests only so that we see this backbone works. After test inclusions,
containerization will not be worked on until the image part is done.

The project architecture is NOT established yet. After the design
processes finishes, and the project frame is visible, then it will be
decided.

* Segregating the documentation from code

One of the most important parts is unraveling the mystery of containers
and images. Now, the code does not contribute to anything for us to understand. I segregated this part for this purpose. Without getting to a certain point in this branch, other branches will wait. This branch not only will include the README and TODOs but also the deep diving all the processes that we will go through. Like a surgeon, I plan to operate on todays technologies and how/what they are doing with the best resolution I can get into. After we have a deep understanding of the underlying the technologies, every code will make sense. So I think this branch is to be the starting point.

* Started in Linux Filesystem Deep Dive

I finished the necessary filesystem coverage for Linux.

* Changes/Fixes according to the PRs.

Now, only the changes suggested are performed. After this push, whole
document will be introduced to a text/grammar checker. It will be a
separate commit for clarity's sake.

The video does not work. It will be fixed on github web.

* Grammarly Check + Corrections

It is more useful than I thought.

* Metadata comparison for README.md and /images directory changed

I took an SS. Than colored where they start is convergent and how they are not aligned
to indicate otherwise.  I mentioned this explicitly. Further
improvements can be made.

* Movie embed

At line 43, there was an attempt to embed a video -> it has failed. GitHub seems to have an improvement on this. If you drag and drop the `.mov` file, it uploads it to a cloud location and references it there.

* Delete `tree_withoout_limits.mov`

It does not serve any purpose.

* Subspace addition

I could not help my self.

* minor fixes

typo fixes